# Pool decode buffers when encoding messages using compression