@rvagg 2025-01-28
> consistent dependency with the one used in Lotus

Not really. https://github.com/filecoin-project/lotus/pull/12015 kind of relevant, there's some pressure to not introduce more cgo lock-in if we can help it, particularly for downstream dependents of those components. I'd say unless the perf improvement here is really dramatic maybe let's just stuck with klauspost/compress.