@masih 2025-01-28
> > consistent dependency with the one used in Lotus
> 
> Not really.

Consistent dependency to the same libraries for the same functionality IMO is a desirable. Maybe the answer here is to move Lotus off the DataDog dependency. I know that `klauspost/compress` zstd package has come a long way since its early days in terms of performance.

Maybe the answer is to move Lotus off DataDog dependency?