@masih 2025-01-28
> I'm surprised we saved an allocation on encode and nothing on decode (where we added the pool). What happens if we remove the pool? Does the datadog version just have an additional allocation internally?

pooling makes no difference to the encoding part. The difference in allocation comes from the internals of the dependency libraries. Here are the numbers for `DataDog/zstd` _without_ pooling:

```
BenchmarkCborEncoding
BenchmarkCborEncoding-12    	  436611	      2375 ns/op	   14680 B/op	      10 allocs/op
BenchmarkCborDecoding
BenchmarkCborDecoding-12    	  106036	     11302 ns/op	   14944 B/op	      27 allocs/op
BenchmarkZstdEncoding
BenchmarkZstdEncoding-12    	  258430	      4643 ns/op	   28248 B/op	      11 allocs/op
BenchmarkZstdDecoding
BenchmarkZstdDecoding-12    	   27158	     43543 ns/op	 1063526 B/op	      28 allocs/op
```